Connected the recipepuppy API
Successfully queried the recipepuppy api with an index request. The api responds with a data object and a results key off that object. The results key has a value that is an array of recipe objects. The Keys to access in each object are: title, thumbnail, ingredients, href. Next step is to connect the href as an open in new tab link to view the whole recipe.