Small UI / UX changes based on user testing
Had some great feedback from user testing and made the following changes.
Clearer dropdown for recipes
Edited Welcome msg on home page
Edited the layout of the recipe card for the show request
Small tweeks to alignment
Adjusted the country cuisine names to be capitalized when they are stored in the api.